Authentication with FOSOAuthBundle

+++ src/AppBundle/Controller/UserController.php
@@ -1,14 +1,15 @@
 <?php
-
 namespace AppBundle\Controller;
 
 use AppBundle\Entity\User;
 use AppBundle\Exception\ResourceValidationException;
+use AppBundle\Service\Approver;
 use Doctrine\ORM\EntityManagerInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use FOS\RestBundle\Controller\Annotations as Rest;
 use FOS\RestBundle\Controller\FOSRestController;
 use FOS\RestBundle\Request\ParamFetcherInterface;
+use FOS\RestBundle\View\View as View;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
@@ -27,12 +28,6 @@ class UserController extends FOSRestController
      *     description="Page on demand"
      * )
      * @Rest\QueryParam(
-     *     name="order",
-     *     requirements="asc|desc",
-     *     default="asc",
-     *     description="Sort order (asc or desc)"
-     * )
-     * @Rest\QueryParam(
      *     name="limit",
      *     requirements="\d+",
      *     default="5",
@@ -48,24 +43,22 @@ class UserController extends FOSRestController
      */
     public function listAction(ParamFetcherInterface $paramFetcher, Request $request, EntityManagerInterface $em)
     {
+        $user = $this->get('security.token_storage')->getToken()->getUser();
+        $id=$user->getClient()->getId();
         $pager = $em->getRepository(User::class)->search(
-            $paramFetcher->get('order'),
+            $id,
             $paramFetcher->get('limit'),
             $paramFetcher->get('offset')
         );
-
         $page=($request->get("page"))?$request->get("page"):1;
         $pager->setCurrentPage($page);
-
         $pagerfantaFactory   = new PagerfantaFactory();
         $paginatedCollection = $pagerfantaFactory->createRepresentation(
             $pager,
             new Route('app_user_list', array())
         );
-
         return $paginatedCollection;
     }
-
     /**
      * @Rest\Get(
      *     path = "/users/{id}",
@@ -74,11 +67,21 @@ class UserController extends FOSRestController
      * )
      * @Rest\View(serializerGroups ={"details"})
      */
-    public function showAction(User $user)
+    public function showAction(User $user, Approver $approver)
     {
+        if (empty($user)) {
+            return View::create(['message' => 'User not found'], Response::HTTP_NOT_FOUND);
+        }
+
+        $admin = $this->get('security.token_storage')->getToken()->getUser();
+
+        if (!$approver->isGranted($user, $admin)){
+            return View::create(['message' => 'You are not allowed to access this resource'], Response::HTTP_FORBIDDEN);
+        }
+
+
         return $user;
     }
-
     /**
      * @Rest\Post(
      *     path = "/users",
@@ -94,13 +97,10 @@ class UserController extends FOSRestController
             foreach ($violations as $violation) {
                 $message .= sprintf("Field %s: %s ", $violation->getPropertyPath(), $violation->getMessage());
             }
-
             throw new ResourceValidationException($message);
         }
-
         $em->persist($user);
         $em->flush();
-
         return $this->view(
             $user,
             Response::HTTP_CREATED,
@@ -109,20 +109,29 @@ class UserController extends FOSRestController
             ]
         );
     }
-
     /**
-     * @Rest\View(StatusCode = 202)
+     * @Rest\View(StatusCode = 204)
      * @Rest\Delete(
      *     path = "/users/{id}",
      *     name = "app_user_delete",
      *     requirements = {"id"="\d+"}
      * )
      */
-    public function deleteAction(User $user, EntityManagerInterface $em)
+    public function deleteAction(User $user, EntityManagerInterface $em, Approver $approver)
     {
-        $em->remove($user);
-        $em->flush();
+        if (empty($user)) {
+            return View::create(['message' => 'User not found'], Response::HTTP_NOT_FOUND);
+        }
 
-        return;
+        $admin = $this->get('security.token_storage')->getToken()->getUser();
+
+        if (!$approver->isGranted($user, $admin)){
+            return View::create(['message' => 'You are not allowed to access this resource'], Response::HTTP_FORBIDDEN);
+        }
+
+        if ($user){
+            $em->remove($user);
+            $em->flush();
+        }
     }
 }
